Silence clang
clang's Weverything includes -Wshorten-64-to-32. But size (size_t
which is 64 bit on macOS) was checked to be less than 23 few lines
prior, so the cast as a counter measure is safe.

+++ link.c
@@ -1101,7 +1101,7 @@ static int matchline (regex_t *re, fz_stext_line *line,
                 trail = "...";
             }
             else {
-                isize = size;
+                isize = (int) size;
                 trail = "";
             }
             printd ("msg regexec error '%*s%s'", isize, errbuf, trail);
